Refactor grouping & enrollment into services
Extract assignment grouping and class enrollment functionality into dedicated controllers and services, and move shared helpers into a controllers.helpers module. Added: assignment_grouping_controller.py, class_enrollment_controller.py, controllers/helpers/assignment_access.py, services/assignment_grouping_service.py, and services/class_enrollment_service.py. Updated api.__init__.py to register new blueprints and updated services.__init__.py exports. Removed in-controller grouping and CSV/enrollment logic from assignment_controller.py and class_controller.py, replacing them with calls to get_teacher_managed_assignment and service functions (clear_assignment_groups, serialize/replace group members, csv_to_list, enroll_students_in_course, etc.) to improve separation of concerns and reuse.
