+++ flask_backend/api/services/__init__.py
@@ -1,9 +1,24 @@
 from .course_grade_service import calculate_student_course_total_grade
 from .class_progress_service import build_class_progress_payload
 from .assignment_progress_service import build_assignment_progress_payload
+from .assignment_grouping_service import (
+	build_grouping_student_payload,
+	clear_assignment_groups,
+	get_course_students,
+	replace_group_members,
+	serialize_assignment_groups,
+)
+from .class_enrollment_service import csv_to_list, enroll_students_in_course
 
 __all__ = [
 	"calculate_student_course_total_grade",
 	"build_class_progress_payload",
 	"build_assignment_progress_payload",
+	"clear_assignment_groups",
+	"get_course_students",
+	"serialize_assignment_groups",
+	"replace_group_members",
+	"build_grouping_student_payload",
+	"csv_to_list",
+	"enroll_students_in_course",
 ]
